feat-bo: rename translations folder
